+++ JoysOfEfficiency/JOEMenu.cs
@@ -7,11 +7,14 @@ using StardewValley.Menus;
 using System.Collections.Generic;
 
 using JoysOfEfficiency.OptionsElements;
+using JoysOfEfficiency.Utils;
 
 namespace JoysOfEfficiency
 {
     public class JOEMenu : IClickableMenu
     {
+        private static readonly string[] fpsLocationStringKey = new string[] { "option.tl", "option.bl", "option.br", "option.tr" };
+
         private ModEntry mod;
 
         private ITranslationHelper translation;
@@ -75,7 +78,7 @@ namespace JoysOfEfficiency
                 tab.AddOptionsElement(new ModifiedCheckBox("AutoGate", 9, ModEntry.Conf.AutoGate, OnCheckboxValueChanged));
                 tab.AddOptionsElement(new ModifiedCheckBox("AutoEat", 10, ModEntry.Conf.AutoEat, OnCheckboxValueChanged));
                 tab.AddOptionsElement(new ModifiedCheckBox("AutoHarvest", 11, ModEntry.Conf.AutoHarvest, OnCheckboxValueChanged));
-                tab.AddOptionsElement(new ModifiedCheckBox("ProtectNectarProducingFlower", 25, ModEntry.Conf.ProtectNectarProducingFlower, OnCheckboxValueChanged));
+                tab.AddOptionsElement(new ModifiedCheckBox("ProtectNectarProducingFlower", 25, ModEntry.Conf.ProtectNectarProducingFlower, OnCheckboxValueChanged, i => !ModEntry.Conf.AutoHarvest));
                 tab.AddOptionsElement(new ModifiedCheckBox("AutoDestroyDeadCrops", 12, ModEntry.Conf.AutoDestroyDeadCrops, OnCheckboxValueChanged));
                 tab.AddOptionsElement(new ModifiedCheckBox("AutoRefillWateringCan", 13, ModEntry.Conf.AutoRefillWateringCan, OnCheckboxValueChanged));
                 tab.AddOptionsElement(new ModifiedCheckBox("AutoCollectCollectibles", 14, ModEntry.Conf.AutoCollectCollectibles, OnCheckboxValueChanged));
@@ -88,6 +91,7 @@ namespace JoysOfEfficiency
                 tab.AddOptionsElement(new ModifiedCheckBox("AutoDepositIngredient", 22, ModEntry.Conf.AutoDepositIngredient, OnCheckboxValueChanged));
                 tab.AddOptionsElement(new ModifiedCheckBox("AutoPullMachineResult,", 23, ModEntry.Conf.AutoPullMachineResult, OnCheckboxValueChanged));
                 tab.AddOptionsElement(new ModifiedCheckBox("AutoPetNearbyPets", 24, ModEntry.Conf.AutoPetNearbyPets, OnCheckboxValueChanged));
+                tab.AddOptionsElement(new ModifiedCheckBox("FPSCounter", 26, ModEntry.Conf.FPSCounter, OnCheckboxValueChanged));
                 tabs.Add(tab);
             }
             {
@@ -104,6 +108,7 @@ namespace JoysOfEfficiency
                 tab.AddOptionsElement(new ModifiedSlider("AutoDigRadius", 8, ModEntry.Conf.AutoDigRadius, 1, 3, OnSliderValueChanged, (() => !ModEntry.Conf.AutoDigArtifactSpot)));
                 tab.AddOptionsElement(new ModifiedSlider("MachineRadius", 10, ModEntry.Conf.MachineRadius, 1, 3, OnSliderValueChanged, (() => !(ModEntry.Conf.AutoPullMachineResult || ModEntry.Conf.AutoDepositIngredient))));
                 tab.AddOptionsElement(new ModifiedSlider("AddedSpeedMultiplier", 9, ModEntry.Conf.AddedSpeedMultiplier, 1, 19, OnSliderValueChanged, (() => !ModEntry.Conf.FasterRunningSpeed)));
+                tab.AddOptionsElement(new ModifiedSlider("FPSlocation", 11, ModEntry.Conf.FPSlocation, 0, 3, OnSliderValueChanged, (() => !ModEntry.Conf.FPSCounter), Format));
                 tabs.Add(tab);
             }
             {
@@ -118,6 +123,10 @@ namespace JoysOfEfficiency
         {
             isListening = true;
             this.listener = listener;
+        }
+        private void OnDropdownChanged(int index, int option)
+        {
+            mod.Monitor.Log($"ID:{index} changed to {option}");
         }
         private void OnInputListnerChanged(int index, Keys value)
         {
@@ -159,6 +168,7 @@ namespace JoysOfEfficiency
                 case 23: ModEntry.Conf.AutoPullMachineResult = value; break;
                 case 24: ModEntry.Conf.AutoPetNearbyPets = value; break;
                 case 25: ModEntry.Conf.ProtectNectarProducingFlower = value; break;
+                case 26: ModEntry.Conf.FPSCounter = value; break;
                 default: return;
             }
             mod.WriteConfig();
@@ -209,6 +219,10 @@ namespace JoysOfEfficiency
             {
                 ModEntry.Conf.MachineRadius = value;
             }
+            if(index == 11)
+            {
+                ModEntry.Conf.FPSlocation = value;
+            }
             mod.WriteConfig();
         }
 
@@ -217,6 +231,10 @@ namespace JoysOfEfficiency
             if (id >= 0 && id < 3)
             {
                 return string.Format("{0:f1}", value / 10f);
+            }
+            else if(id == 11)
+            {
+                return mod.Helper.Translation.Get(fpsLocationStringKey[value]);
             }
             return value + "";
         }
@@ -252,14 +270,14 @@ namespace JoysOfEfficiency
             int x = 16, y = 16;
 
 
-            drawTextureBox(b, tabEnabled.Left, tabEnabled.Top, tabEnabled.Width, tabEnabled.Height, Color.White * (tabIndex != 0 ? 1.0f : 0.6f));
-            b.DrawString(Game1.smallFont, tabEnabledString, new Vector2(tabEnabled.Left + 16, tabEnabled.Top + (tabEnabled.Height - font.MeasureString(tabEnabledString).Y) / 2), Color.Black * (tabIndex != 0 ? 1.0f : 0.6f));
+            drawTextureBox(b, tabEnabled.Left, tabEnabled.Top, tabEnabled.Width, tabEnabled.Height, Color.White * (tabIndex == 0 ? 1.0f : 0.6f));
+            b.DrawString(Game1.smallFont, tabEnabledString, new Vector2(tabEnabled.Left + 16, tabEnabled.Top + (tabEnabled.Height - font.MeasureString(tabEnabledString).Y) / 2), Color.Black * (tabIndex == 0 ? 1.0f : 0.6f));
 
-            drawTextureBox(b, tabSliders.Left, tabSliders.Top, tabSliders.Width, tabSliders.Height, Color.White * (tabIndex != 1 ? 1.0f : 0.6f));
-            b.DrawString(Game1.smallFont, tabSlidersString, new Vector2(tabSliders.Left + 16, tabSliders.Top + (tabSliders.Height - font.MeasureString(tabSlidersString).Y) / 2), Color.Black * (tabIndex != 1 ? 1.0f : 0.6f));
+            drawTextureBox(b, tabSliders.Left, tabSliders.Top, tabSliders.Width, tabSliders.Height, Color.White * (tabIndex == 1 ? 1.0f : 0.6f));
+            b.DrawString(Game1.smallFont, tabSlidersString, new Vector2(tabSliders.Left + 16, tabSliders.Top + (tabSliders.Height - font.MeasureString(tabSlidersString).Y) / 2), Color.Black * (tabIndex == 1 ? 1.0f : 0.6f));
 
-            drawTextureBox(b, tabControls.Left, tabControls.Top, tabControls.Width, tabControls.Height, Color.White * (tabIndex != 2 ? 1.0f : 0.6f));
-            b.DrawString(Game1.smallFont, tabControlsString, new Vector2(tabControls.Left + 16, tabControls.Top + (tabControls.Height - font.MeasureString(tabControlsString).Y) / 2), Color.Black * (tabIndex != 2 ? 1.0f : 0.6f));
+            drawTextureBox(b, tabControls.Left, tabControls.Top, tabControls.Width, tabControls.Height, Color.White * (tabIndex == 2 ? 1.0f : 0.6f));
+            b.DrawString(Game1.smallFont, tabControlsString, new Vector2(tabControls.Left + 16, tabControls.Top + (tabControls.Height - font.MeasureString(tabControlsString).Y) / 2), Color.Black * (tabIndex == 2 ? 1.0f : 0.6f));
 
             drawTextureBox(b, Game1.menuTexture, new Rectangle(0, 256, 60, 60), xPositionOnScreen, yPositionOnScreen, this.width, this.height, Color.White, 1.0f, false);
             base.draw(b);
@@ -288,11 +306,8 @@ namespace JoysOfEfficiency
                 drawTextureBox(b, (Game1.viewport.Width - size.X) / 2, (Game1.viewport.Height - size.Y) / 2, size.X, size.Y, Color.White);
                 listener.DrawStrings(b, (Game1.viewport.Width - size.X) / 2, (Game1.viewport.Height - size.Y) / 2);
             }
-
-            if (!Game1.options.hardwareCursor)
-            {
-                b.Draw(Game1.mouseCursors, new Vector2(Game1.getOldMouseX(), Game1.getOldMouseY()), Game1.getSourceRectForStandardTileSheet(Game1.mouseCursors, Game1.options.gamepadControls ? 44 : 0, 16, 16), Color.White, 0f, Vector2.Zero, Game1.pixelZoom + Game1.dialogueButtonScale / 150f, SpriteEffects.None, 1f);
-            }
+            
+            Util.DrawCursor();
         }
 
         public override void performHoverAction(int x, int y)
@@ -418,9 +433,10 @@ namespace JoysOfEfficiency
             for (int i = firstIndex; i < menuElements.Count; i++)
             {
                 OptionsElement element = menuElements[i];
-                if (y + element.bounds.Height < height)
+                int hElem = element is ModifiedSlider ? element.bounds.Height + 4 : element.bounds.Height;
+                if (y + hElem < height)
                 {
-                    y += element.bounds.Height + 16;
+                    y += hElem + 16;
                     elements.Add(element);
                 }
             }
@@ -434,9 +450,10 @@ namespace JoysOfEfficiency
             for (int i = firstIndex; i < menuElements.Count; i++)
             {
                 OptionsElement element = menuElements[i];
-                if (y + element.bounds.Height < height)
+                int hElem = element is ModifiedSlider ? element.bounds.Height + 4 : element.bounds.Height;
+                if (y + hElem < height)
                 {
-                    y += element.bounds.Height + 16;
+                    y += hElem + 16;
                 }
                 else
                 {